Improve compatibility with pathlib by implementing more common methods (#230)
* Added absolute, resolve and relative_to implementations

* Added tests interacting with PurePosixPaths

* Added is_absolute and is_relative_to as well

* Accept read_text parameters

* Added exist_ok paramter to touch

* Added missing_ok paramter to unlink

* Fixed formatting (make format)

* Update history.md & readme.md

* Fix relative_to on different providers (+ some naming changes

* Reformat

* Update history.md and bump version number to 0.9

* Update dev requirements for docs

* Fix readme list order

* Added test for relative_to on different provider cloud paths

* Remove unused import

* Updated relative_to different cloud test

+++ cloudpathlib/local/localclient.py
@@ -96,8 +96,11 @@ class LocalClient(Client):
             shutil.copy(self._cloud_path_to_local(src), self._cloud_path_to_local(dst))
         return dst
 
-    def _remove(self, cloud_path: "LocalPath") -> None:
+    def _remove(self, cloud_path: "LocalPath", missing_ok: bool = True) -> None:
         local_storage_path = self._cloud_path_to_local(cloud_path)
+        if not missing_ok and not local_storage_path.exists():
+            raise FileNotFoundError(f"File does not exist: {cloud_path}")
+
         if local_storage_path.is_file():
             local_storage_path.unlink()
         elif local_storage_path.is_dir():
@@ -121,8 +124,10 @@ class LocalClient(Client):
             )
         )
 
-    def _touch(self, cloud_path: "LocalPath") -> None:
+    def _touch(self, cloud_path: "LocalPath", exist_ok: bool = True) -> None:
         local_storage_path = self._cloud_path_to_local(cloud_path)
+        if local_storage_path.exists() and not exist_ok:
+            raise FileExistsError(f"File exists: {cloud_path}")
         local_storage_path.parent.mkdir(exist_ok=True, parents=True)
         local_storage_path.touch()
